refactor(auth): isola lógica de autenticação no domínio auth
